+++ spotify/__init__.py
@@ -14,6 +14,7 @@ from _spotify import PlaylistContainer
 from _spotify import AlbumBrowser
 from _spotify import ArtistBrowser
 from _spotify import Image
+from _spotify import User
 
 from _spotify import api_version
 from _spotify import connect