+++ spotify/constant.py
@@ -1,0 +1,8 @@
+# encoding: utf-8
+
+# User relation types
+from _spotify import                \
+RELATION_TYPE_UNKNOWN,              \
+RELATION_TYPE_NONE,                 \
+RELATION_TYPE_UNIDIRECTIONAL,       \
+RELATION_TYPE_BIDIRECTIONAL