+++ src/session.h
@@ -1,7 +1,10 @@
 typedef struct {
-    PyObject_HEAD sp_session *_session;
+    PyObject_HEAD
+    sp_session *_session;
     PyObject *client;
 } Session;
+
+sp_session *g_session;
 
 extern PyTypeObject SessionType;
 extern void session_init(PyObject *m);