+++ src/mockmodule.c
@@ -1394,7 +1394,6 @@ mock_albumbrowse(PyObject *self, PyObject *args, PyObject *kwds)
     ab = (AlbumBrowser *) PyObject_Call((PyObject *)&AlbumBrowserType,
                                         new_args, NULL);
     ab->_browser->loaded = loaded;
-    Py_INCREF(ab);
     return (PyObject *)ab;
 }
 
@@ -1438,7 +1437,6 @@ mock_artistbrowse(PyObject *self, PyObject *args, PyObject *kwds)
     ab = (ArtistBrowser *) PyObject_Call((PyObject *)&ArtistBrowserType,
                                          new_args, NULL);
     ab->_browser->loaded = loaded;
-    Py_INCREF(ab);
     return (PyObject *)ab;
 }