+++ tests/test_albumbrowser.py
@@ -2,7 +2,6 @@ import unittest
 from spotify._mockspotify import mock_albumbrowse, mock_album, mock_artist
 from spotify._mockspotify import mock_session
 from spotify import Album
-from tests import SkipTest
 
 callback_called = False
 callback_userdata = None
@@ -13,7 +12,9 @@ class TestAlbumbrowser(unittest.TestCase):
     album = mock_album("foo0", mock_artist("bar0", 1), 2006,
                                "01234567890123456789", Album.ALBUM, 1, 1)
 
-    def callback(browser, userdata):
+    def callback(self, browser, userdata):
+        global callback_called
+        global callback_userdata
         callback_called = True
         callback_userdata = userdata
 
@@ -33,9 +34,10 @@ class TestAlbumbrowser(unittest.TestCase):
         assert browser[2].name() == 'baz'
 
     def test_callback(self):
-        raise SkipTest
+        global callback_called
+        global callback_userdata
         callback_called = False
         browser = mock_albumbrowse(self.session, self.album, 0, self.callback,
                                    self)
-        assert callback_called
-        assert userdata is self
+        self.assertTrue(callback_called)
+        self.assertEqual(callback_userdata, self)