+++ docs/changes.rst
@@ -5,6 +5,7 @@ Changes
 v1.4 (in development)
 =====================
 
+- User handling
 - New Playlist method: rename.
 
 v1.3 (2011-06-11)