+++ docs/changes.rst
@@ -17,6 +17,10 @@ Bug fixes:
 
 - log_message callbacks used ``str`` in place of ``unicode``.
 - argument errors were unchecked in Session.search
+- Fix fail on valid error at image creation (jkp)
+- Keep compatibility with Python 2.5 (jkp)
+- Callbacks given at artist/album browser creation are now called by pyspotify
+  (jkp)
 
 v1.3 (2011-06-11)
 =================